Bring forward contributors page content from previous PR
This brings in a contributors page for the Project Ricotta
contributors, as envisioned in https://github.com/Lasagna-Love-Portal/project-ricotta/pull/20.
This PR adds the page as a new contributors page with some very
minimal styling and with an initial list of known contributors.

This does not bring forward the general styling CSS from that PR,
as we'll be developing some site-wide CSS and styling later.

Includes very basic link from welcome landing page.