Initial JWT refresh token usage implementation
Initial logic added to store access and refresh tokens that will
be returned by bechamel's login API, and to use the refresh token
to request updated access/refresh token pair from bechamel when
an HTTP 401 is returned from the /profile API in getCurrentUserProfile.

Some refactoring and better error handling is due for this.